Adds a basic script for generating API docs from the `mediapipe` Python package
This is pretty bare-bones for now, but I need something to start wiring up the rest of the automation. Plus this should be easy enough for anyone to riff on once it's in place.

I couldn't find a great location within the existing directory structure for this, so LMK if it should be re-homed.

PiperOrigin-RevId: 484426543