Add itemized loop calculators
Adds Begin/EndItemLoopCalculator, which work similar to Begin/EndLoopCalculator, but take individual ITEMs instead of an iterable. This has the advantage that inputs/outputs don't have to be (de-)vectorized. Vectorization can be a problem for items that are not copyable, such as Tensors. So far, looping over a collection of tensors was only possible if the collection was consumable, i.e. not referenced anywhere else. This change removes the need of vectorization, but has the restriction that the the size of the set of items needs to be fixed and known.

PiperOrigin-RevId: 625756766