Remove Resources::ReadContents & add Resource::TryReleaseAsString.
It's cumbersome to implement both Resources::ReadContents and Resources::Get, also the idea to switch to Resource::Get completely, so the best way (instead of deprecating) is to just remove ReadContents as it's quite new.

The alternative is added if `std::string` is needed (for performance/apis reason) - user still can try to release/consume resource as string, but if it's impossible (e.g. some embedded resource), they will need to fallback to a copy using `ToStringView`.

PiperOrigin-RevId: 675594664