Add tools for logging Tensors, ImageFrames and cv::Mats
Adds debug::LogTensor/Image/Mat that print out a terminal-friendly version of the object. For terminals that support truecolor ($COLORTERM == "truecolor"), prints a low-res (<120 pixels in x and y) version of the image. For other terminals, prints an ASCII art version. LogImage and LogMat tools work for all formats and channel counts. LogTensor assumes float tensors of size [1,h,w,c].

This tooling is super useful for debugging graph issues, e.g. to verify if images and tensors in different stages of a pipeline are oriented correctly (not upside-down).

PiperOrigin-RevId: 738821402