Add a test parameter to ignore pixels above diff limit
Adds an optional parameter to ImageFrameComparisonOptions to ignore a few pixels above the difference limits when comparing images. This is useful in situations where code that is not under our control behaves slightly non-deterministically and produces a few off-pixels that should be ignored. Defaults to 0, so there should be no difference to the current behavior if the parameter is not set.

PiperOrigin-RevId: 774823518